pci-id file hierarchy implementation, local, distribution, repository

+++ GPUmodules/PCImodule.py
@@ -21,7 +21,7 @@ __copyright__ = 'Copyright (C) 2019 RueiKe'
 __credits__ = ['']
 __license__ = 'GNU General Public License'
 __program_name__ = 'amdgpu-utils'
-__version__ = 'v2.6.0'
+__version__ = 'v2.7.0'
 __maintainer__ = 'RueiKe'
 __status__ = 'Stable Release'
 
@@ -33,21 +33,47 @@ from pathlib import Path
 import shutil
 import urllib.request
 
+try:
+    from GPUmodules import env
+except:
+    import env
+
 
 class PCI_ID:
     def __init__(self, file_name='amd_pci_id.txt'):
         self.pciid_url = 'https://pci-ids.ucw.cz/v2.2/pci.ids'
-        self.pciid_file = 'pci.ids'
-        self.amdgpu_utils_file = os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name)
-        try:
-            self.pci_id_file_ptr = open(self.amdgpu_utils_file, 'r')
-        except FileNotFoundError:
-            print('File [%s] not found. Exiting...' %
-                  os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name))
-            sys.exit(-1)
-        except OSError:
-            print('Can not open [%s] to read. Exiting...' %
-                  os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name))
+        self.pciid_file = 'pci.ids'    # base of name for downloaded file
+
+        # Possible locations of PCI ID files
+        self.pciid_file_local = os.path.join(env.gut_const.config_dir, file_name)
+        self.pciid_file_distribution = os.path.join(env.gut_const.dist_share, file_name)
+        self.pciid_file_repository = os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name)
+
+        # Details on PCI-ID file in use
+        self.file_open_status = False
+        self.amdgpu_utils_file = None
+        self.pci_id_file_ptr = None
+
+        for try_filename in [self.pciid_file_local, self.pciid_file_distribution, self.pciid_file_repository]:
+            if env.gut_const.DEBUG: print('Trying pci-id file [{}].'.format(try_filename))
+            self.amdgpu_utils_file = try_filename
+            if not os.path.isfile(self.amdgpu_utils_file):
+                continue
+            try:
+                self.pci_id_file_ptr = open(self.amdgpu_utils_file, 'r')
+            except FileNotFoundError:
+                print('File [%s] not found.' %
+                      os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name))
+                continue
+            except OSError:
+                print('Can not open [%s] to read.' %
+                      os.path.join(os.path.dirname(str(Path(__file__).resolve())), file_name))
+                continue
+            finally:
+                self.file_open_status = True
+                return
+        if not self.file_open_status:
+            print('Can not find a valid PCI-ID file: [%s]. Exiting...' % file_name)
             sys.exit(-1)
 
     def get_pciid_version(self, filename=''):
@@ -90,16 +116,30 @@ class PCI_ID:
         with urllib.request.urlopen(self.pciid_url) as response, open(file_name, 'wb') as out_file:
             shutil.copyfileobj(response, out_file)
 
-        return file_name
+        if os.path.isfile(file_name):
+            return file_name
+        else:
+            return None
 
     def update_pci_id(self, in_file_name):
+        if not os.path.isdir(env.gut_const.config_dir):
+            os.mkdir(env.gut_const.config_dir)
+        if not os.path.isdir(env.gut_const.config_dir):
+            print('ERROR: could not create amdgpu-utils config directory: [{}]'.format(env.gut_const.config_dir))
+            sys.exit(-1)
         self.pci_id_file_ptr.close()
-        self.extract_vendor_from_pci_id('0x1002', in_file_name, self.amdgpu_utils_file)
+        # self.extract_vendor_from_pci_id('0x1002', in_file_name, self.amdgpu_utils_file)
+        self.extract_vendor_from_pci_id('0x1002', in_file_name, self.pciid_file_local)
+        self.file_open_status = False
+        self.amdgpu_utils_file = self.pciid_file_local
+        self.pci_id_file_ptr = None
         return 0
 
     @staticmethod
     def extract_vendor_from_pci_id(vendor, in_file_name, out_file_name=''):
         """ For a given vendor id, extract all relevant entries.
+
+            Write extracted entries to out_file_name if specified.
         """
         try:
             in_file_ptr = open(in_file_name, 'r')
@@ -113,7 +153,8 @@ class PCI_ID:
         if out_file_name == '':
             file_ptr = sys.stdout
         else:
-            shutil.copy2(out_file_name, out_file_name + datetime.utcnow().strftime('%m%d_%H%M%S'))
+            if os.path.isfile(out_file_name):
+                shutil.copy2(out_file_name, out_file_name + datetime.utcnow().strftime('%m%d_%H%M%S'))
             try:
                 file_ptr = open(out_file_name, 'w')
             except OSError:
